#364 Move footnote tests to their own class and folder.
