+++ openhtmltopdf-examples/src/main/resources/visualtest/html/footnote/issue-364-footnotes-blocks.html
